backend/requirements.txt
fastapi==0.92.0
pydantic==1.10.5
python-dotenv==0.21.1
python-socketio==5.7.2
pytz==2023.3
pymongo~=3.12.0
beanie~=1.21.0
redis==4.5.1
rich==13.4.2
starlette==0.25.0
uvicorn==0.20.0